feat: The real test: Does it refactor?
