+++ src/components/__tests__/VTU/Navbar.spec.js
@@ -1,0 +1,25 @@
+import { mount } from '@vue/test-utils'
+import Navbar from '../../Navbar.vue'
+
+describe('Navbar', () => {
+  function navbarFactory(props) {
+    return mount(Navbar, {
+      props
+    })
+  }
+
+  it('shows login authenticated is true', () => {
+    const wrapper = navbarFactory({ authenticated: true })
+    expect(wrapper.html()).toContain('Logout')
+  })
+
+  it('shows logout by default', () => {
+    const wrapper = navbarFactory()
+    expect(wrapper.find('a').text()).toBe('Login')
+  })
+
+  it('shows login when authenticated is false', () => {
+    const wrapper = navbarFactory({ authenticated: false })
+    expect(wrapper.find('a').text()).toBe('Login')
+  })
+})
